コメント機能をリファクタリング
